generate self healing move interface
