+++ src/battle/effect/generic/EffectInterfaces.java
@@ -124,6 +124,18 @@ public final class EffectInterfaces {
 		}
 	}
 
+	public interface SelfHealingMove {
+		double getHealFraction(Battle b, ActivePokemon victim);
+
+		default void heal(Battle b, ActivePokemon victim) {
+			// Heal yourself!
+			victim.healHealthFraction(this.getHealFraction(b, victim));
+			
+			// TODO: Make sure the message is set up correctly for the hp change
+			Messages.add(new MessageUpdate(victim.getName() + "'s health was restored!").updatePokemon(b, victim));
+		}
+	}
+
 	public interface PhysicalContactEffect {
 
 		// b: The current battle
